pyproject.toml
[tool.poetry]
name = "lmanage"
version = "0.3.15"
description = "LManage is a collection of useful tools for Looker admins to help curate and cleanup content and it's associated source LookML."
keywords = ["keyword", "another_keyword"]
authors = ["hselbie <hselbie@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/looker-open-source/lmanage"
repository = "https://github.com/looker-open-source/lmanage"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
looker-sdk = "^23.8.1"
poetry = "^1.4.2"
rich = "^13.4.2"
click = "^8.1.3"
ruamel-yaml = "^0.17.32"
tqdm = "^4.65.0"
tenacity = "^8.2.2"
yaspin = "^2.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
lmanage = 'lmanage.cli:lmanage'
